Start an OpenMDAO-related refactor
